backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
elasticsearch>=8.15.0
python-dotenv>=1.0.1
httpx>=0.27.0
pydantic>=2.9.0
email-validator>=2.2.0
